+++ refs.c
@@ -29,25 +29,25 @@ uint64_t search_refs(char* name, RefType type, bool *found) {
 
 void resolve_refs() {
     // for (int32_t i = 0; i < refrences_size; i++) {
-    //     printf("Name: %s | FuncAddr: %d | CodeAddr: %d | Type: %d\n", refrences[i].name, refrences[i].func_idx, refrences[i].addr, refrences[i].type);
+    //     printf("Name: %s | FuncAddr: %lld | CodeAddr: %lld | Type: %d\n", refrences[i].name, refrences[i].func_idx, refrences[i].addr, (int64_t) refrences[i].type);
     // }
     // printf("======\n");
     // for (int32_t i = 0; i < unresolved_refrences_size; i++) {
-    //     printf("Name: %s | FuncAddr: %d | CodeAddr: %d | Type: %d\n", unresolved_refrences[i].name, unresolved_refrences[i].func_idx, unresolved_refrences[i].addr, unresolved_refrences[i].type);
+    //     printf("Name: %s | FuncAddr: %lld | CodeAddr: %lld | Type: %d\n", unresolved_refrences[i].name, unresolved_refrences[i].func_idx, unresolved_refrences[i].addr, (int64_t) unresolved_refrences[i].type);
     // }
     bool found;
     uint64_t found_addr;
-    int64_t relative_addr;
+    uint8_t relative_addr;
     for (uint64_t i = 0; i < unresolved_refrences_size; i++) {
         found_addr = search_refs(unresolved_refrences[i].name, unresolved_refrences[i].type, &found);
         if (found) {
             if (unresolved_refrences[i].type == JMP_REF) {
-                relative_addr = refrences[found_addr].addr - (unresolved_refrences[i].addr-1);
-                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(int16_t));
+                relative_addr = (uint8_t) refrences[found_addr].addr - (unresolved_refrences[i].addr-1);
+                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(int8_t));
             }
             else if (unresolved_refrences[i].type == FUNC_REF) {
-                relative_addr = refrences[found_addr].func_idx;
-                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(int16_t));
+                relative_addr = (int8_t) refrences[found_addr].func_idx;
+                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(int8_t));
             }
             else {
                 printf("unknown ref typen\n"); exit(1);
